refactor(native): organize engine sources
