+++ native/CMakeLists.txt
@@ -10,14 +10,14 @@ endif()
 
 option(PROTON_WITH_ENGINE "Build Proton with the native browser engine linked" OFF)
 set(PROTON_ENGINE_ROOT "" CACHE PATH "Native browser engine runtime/SDK root")
-set(PROTON_ENGINE_SOURCE src/proton_engine_none.c)
+set(PROTON_ENGINE_SOURCE src/engine/proton_engine_none.c)
 
 if(PROTON_WITH_ENGINE)
   if(WIN32)
-    set(PROTON_ENGINE_SOURCE src/proton_engine_cef_win.c)
+    set(PROTON_ENGINE_SOURCE src/engine/cef_win/proton_engine_cef_win.c)
   elseif(APPLE)
     enable_language(OBJC)
-    set(PROTON_ENGINE_SOURCE src/proton_engine_cef_mac.m)
+    set(PROTON_ENGINE_SOURCE src/engine/cef_mac/proton_engine_cef_mac.m)
   else()
     message(FATAL_ERROR "PROTON_WITH_ENGINE is not wired for this platform")
   endif()
